+++ CHANGELOG.md
@@ -2,10 +2,21 @@
 
 Notable changes to the project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [0.0.3] - 30th November 2023
 
-## 0.0.1 - 28th November 2023
+### Added
+
+- Added stylesheet creation to component
+- Added configuration toggle for stylesheet creation
+- Added TypeScript/JavaScript detection
+
+## [0.0.2] - 29th November 2023
+
+### Added
+
+- Create Page command
+
+## [0.0.1] - 28th November 2023
 
 ### Added